fix priority queue
