+++ trinity/buffer/ray_wrapper.py
@@ -1,6 +1,9 @@
+import asyncio
 import json
 import os
 import time
+from collections import deque
+from copy import deepcopy
 from typing import List, Optional
 
 import ray
@@ -9,10 +12,12 @@ from sqlalchemy.exc import OperationalError
 from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import NullPool
 
+from trinity.buffer.queue import QueueBuffer
 from trinity.buffer.schema import Base, create_dynamic_table
 from trinity.buffer.utils import default_storage_path, retry_session
+from trinity.buffer.writer.file_writer import JSONWriter
 from trinity.common.config import BufferConfig, StorageConfig
-from trinity.common.constants import ReadStrategy
+from trinity.common.constants import ReadStrategy, StorageType
 from trinity.common.experience import Experience
 from trinity.common.workflows import Task
 from trinity.utils.log import get_logger
@@ -199,3 +204,97 @@ class FileWrapper:
         if self.ref_count <= 0:
             self.file.close()
         return self.ref_count
+
+
+def is_database_url(path: str) -> bool:
+    return any(path.startswith(prefix) for prefix in ["sqlite:///", "postgresql://", "mysql://"])
+
+
+def is_json_file(path: str) -> bool:
+    return path.endswith(".json") or path.endswith(".jsonl")
+
+
+class QueueWrapper:
+    """An wrapper of a async queue."""
+
+    def __init__(self, storage_config: StorageConfig, config: BufferConfig) -> None:
+        self.logger = get_logger(__name__)
+        self.config = config
+        self.capacity = storage_config.capacity
+        self.queue = QueueBuffer.get_queue(storage_config, config)
+        st_config = deepcopy(storage_config)
+        st_config.wrap_in_ray = False
+        if st_config.path is not None:
+            if is_database_url(st_config.path):
+                from trinity.buffer.writer.sql_writer import SQLWriter
+
+                st_config.storage_type = StorageType.SQL
+                self.writer = SQLWriter(st_config, self.config)
+            elif is_json_file(st_config.path):
+                st_config.storage_type = StorageType.FILE
+                self.writer = JSONWriter(st_config, self.config)
+            else:
+                self.logger.warning("Unknown supported storage path: %s", st_config.path)
+                self.writer = None
+        else:
+            st_config.storage_type = StorageType.FILE
+            self.writer = JSONWriter(st_config, self.config)
+        self.logger.warning(f"Save experiences in {st_config.path}.")
+        self.ref_count = 0
+        self.exp_pool = deque()  # A pool to store experiences
+        self.closed = False
+
+    async def acquire(self) -> int:
+        self.ref_count += 1
+        return self.ref_count
+
+    async def release(self) -> int:
+        """Release the queue."""
+        self.ref_count -= 1
+        if self.ref_count <= 0:
+            self.queue.close()
+            await self.writer.release()
+        return self.ref_count
+
+    def length(self) -> int:
+        """The length of the queue."""
+        return self.queue.qsize()
+
+    async def put_batch(self, exp_list: List) -> None:
+        """Put batch of experience."""
+        await self.queue.put(exp_list)
+        if self.writer is not None:
+            self.writer.write(exp_list)
+
+    async def get_batch(self, batch_size: int, timeout: float) -> List:
+        """Get batch of experience."""
+        while len(self.exp_pool) < batch_size:
+            if self.queue.stopped():
+                raise StopAsyncIteration("Queue is closed and no more items to get.")
+            try:
+                exp_list = await asyncio.wait_for(self.queue.get(), timeout=timeout)
+            except asyncio.TimeoutError:
+                self.logger.error(
+                    f"Timeout when waiting for experience, only get {len(self.exp_pool)} experiences.\n"
+                    "This phenomenon is usually caused by the workflow not returning enough "
+                    "experiences or running timeout. Please check your workflow implementation."
+                )
+                result = list(self.exp_pool)
+                self.exp_pool.clear()
+                return result
+            self.exp_pool.extend(exp_list)
+        batch = [self.exp_pool.popleft() for _ in range(batch_size)]
+        return batch
+
+    @classmethod
+    def get_wrapper(cls, storage_config: StorageConfig, config: BufferConfig):
+        """Get the queue actor."""
+        return (
+            ray.remote(cls)
+            .options(
+                name=f"queue-{storage_config.name}",
+                namespace=storage_config.ray_namespace or ray.get_runtime_context().namespace,
+                get_if_exists=True,
+            )
+            .remote(storage_config, config)
+        )
